# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KIPS School Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: date
          in: query
          schema:
            type: ["string", "null"]
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
            enum: [student, staff, null]
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: markAttendance
      tags: [attendance]
      summary: Mark attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Attendance marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /exams:
    get:
      operationId: listExams
      tags: [exams]
      summary: List exams
      responses:
        "200":
          description: List of exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
    post:
      operationId: createExam
      tags: [exams]
      summary: Create exam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamInput"
      responses:
        "201":
          description: Exam created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /exams/{id}/results:
    get:
      operationId: getExamResults
      tags: [exams]
      summary: Get exam results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamResult"
    post:
      operationId: saveExamResult
      tags: [exams]
      summary: Save exam result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamResultInput"
      responses:
        "201":
          description: Result saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResult"
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff members
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Staff"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Staff created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
  /staff/{id}:
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update staff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
  /salaries:
    get:
      operationId: listSalaries
      tags: [staff]
      summary: List salary records
      parameters:
        - name: staffId
          in: query
          schema:
            type: ["integer", "null"]
        - name: month
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [paid, unpaid, null]
      responses:
        "200":
          description: Salary records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Salary"
    post:
      operationId: createSalary
      tags: [staff]
      summary: Create salary record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalaryInput"
      responses:
        "201":
          description: Salary record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salary"
  /salaries/{id}/pay:
    post:
      operationId: paySalary
      tags: [staff]
      summary: Mark salary as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Salary paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salary"
  /accounts/income:
    get:
      operationId: listIncome
      tags: [accounts]
      summary: List income records
      parameters:
        - name: month
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Income records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountEntry"
    post:
      operationId: createIncome
      tags: [accounts]
      summary: Create income record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountEntryInput"
      responses:
        "201":
          description: Income created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountEntry"
  /accounts/expenses:
    get:
      operationId: listExpenses
      tags: [accounts]
      summary: List expense records
      parameters:
        - name: month
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Expense records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountEntry"
    post:
      operationId: createExpense
      tags: [accounts]
      summary: Create expense record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountEntryInput"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountEntry"
  /accounts/summary:
    get:
      operationId: getAccountSummary
      tags: [accounts]
      summary: Get income/expense summary
      parameters:
        - name: month
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Account summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountSummary"
...
components:
  schemas:
...
    Attendance:
      type: object
      required: [id, date, type, status, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: ["integer", "null"]
        staffId:
          type: ["integer", "null"]
        personName:
          type: ["string", "null"]
        className:
          type: ["string", "null"]
        date:
          type: string
        type:
          type: string
          enum: [student, staff]
        status:
          type: string
          enum: [present, absent, late, leave]
        createdAt:
          type: string
    AttendanceInput:
      type: object
      required: [date, type, status]
      properties:
        studentId:
          type: ["integer", "null"]
        staffId:
          type: ["integer", "null"]
        date:
          type: string
        type:
          type: string
          enum: [student, staff]
        status:
          type: string
          enum: [present, absent, late, leave]
    Exam:
      type: object
      required: [id, name, classId, subject, examDate, totalMarks, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        classId:
          type: integer
        className:
          type: ["string", "null"]
        subject:
          type: string
        examDate:
          type: string
        totalMarks:
          type: number
        passingMarks:
          type: number
        createdAt:
          type: string
    ExamInput:
      type: object
      required: [name, classId, subject, examDate, totalMarks]
      properties:
        name:
          type: string
        classId:
          type: integer
        subject:
          type: string
        examDate:
          type: string
        totalMarks:
          type: number
        passingMarks:
          type: number
    ExamResult:
      type: object
      required: [id, examId, studentId, marksObtained, grade, createdAt]
      properties:
        id:
          type: integer
        examId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        marksObtained:
          type: number
        grade:
          type: string
        position:
          type: ["integer", "null"]
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
    ExamResultInput:
      type: object
      required: [studentId, marksObtained]
      properties:
        studentId:
          type: integer
        marksObtained:
          type: number
        grade:
          type: string
        position:
          type: ["integer", "null"]
        remarks:
          type: string
    Staff:
      type: object
      required: [id, name, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
          enum: [teacher, admin, accountant, support]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        salary:
          type: ["number", "null"]
        joinDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        username:
          type: ["string", "null"]
        createdAt:
          type: string
    StaffInput:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
        role:
          type: string
          enum: [teacher, admin, accountant, support]
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        subject:
          type: string
        salary:
          type: number
        joinDate:
          type: string
        status:
          type: string
          enum: [active, inactive]
    StaffUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        subject:
          type: string
        salary:
          type: number
        status:
          type: string
          enum: [active, inactive]
    Salary:
      type: object
      required: [id, staffId, amount, month, status, createdAt]
      properties:
        id:
          type: integer
        staffId:
          type: integer
        staffName:
          type: ["string", "null"]
        amount:
          type: number
        month:
          type: string
        paidDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [paid, unpaid]
        createdAt:
          type: string
    SalaryInput:
      type: object
      required: [staffId, amount, month]
      properties:
        staffId:
          type: integer
        amount:
          type: number
        month:
          type: string
    AccountEntry:
      type: object
      required: [id, type, amount, category, date, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        category:
          type: string
        description:
          type: string
        date:
          type: string
        createdAt:
          type: string
    AccountEntryInput:
      type: object
      required: [amount, category, description, date]
      properties:
        amount:
          type: number
        category:
          type: string
        description:
          type: string
        date:
          type: string
    AccountSummary:
      type: object
      required: [totalIncome, totalExpenses, netProfit]
      properties:
        totalIncome:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        month:
          type: ["string", "null"]
...