# lib/api-zod/src/generated/types/staffInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KIPS School Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { StaffInputRole } from './staffInputRole';
import type { StaffInputStatus } from './staffInputStatus';

export interface StaffInput {
  name: string;
  role: StaffInputRole;
  phone?: string;
  email?: string;
  address?: string;
  subject?: string;
  salary?: number;
  joinDate?: string;
  status?: StaffInputStatus;
}